requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
discord==1.7.3
discord.py==1.7.3
Django==3.2.9
djangorestframework==3.12.4
idna==3.3
multidict==5.2.0
python-dotenv==0.19.1
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
yarl==1.7.2
